pyproject.toml
[project]
name = "ai-sales-mcp-demo"
version = "0.1.0"
description = "AI Sales MCP Demo"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "pandas>=2.0.0",
    "sqlite-utils>=3.35",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "streamlit-lottie==0.0.5",
    "streamlit-option-menu==0.3.6",
    "streamlit-extras==0.3.6",
    "colorama>=0.4.6",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.177.0",
    "python-docx>=1.2.0",
    "psutil>=7.0.0",
    "spacy>=3.8.7",
    "pyjwt>=2.10.1",
    "bcrypt>=4.3.0",
    "openai>=1.98.0",
    "xlsxwriter>=3.2.5",
    "aiohttp>=3.12.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["servers", "client", "tests"]
